LPS-53627 extract locales functionality to its own module
