+++ modules/apps/journal/journal-content-asset-addon-entries/journal-content-locales-asset-addon-entry/src/META-INF/resources/locales.jsp
@@ -1,0 +1,57 @@
+<%--
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+--%>
+
+<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
+
+<%@ taglib uri="http://java.sun.com/portlet_2_0" prefix="portlet" %>
+
+<%@ taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
+
+<%@ page contentType="text/html; charset=UTF-8" %>
+
+<%@ page import="com.liferay.portal.kernel.language.LanguageUtil" %>
+<%@ page import="com.liferay.portal.kernel.util.Constants" %>
+<%@ page import="com.liferay.portal.kernel.util.ParamUtil" %>
+<%@ page import="com.liferay.portal.kernel.util.WebKeys" %>
+<%@ page import="com.liferay.portlet.PortletURLUtil" %>
+<%@ page import="com.liferay.portlet.journal.model.JournalArticleDisplay" %>
+
+<%@ page import="javax.portlet.PortletURL" %>
+
+<portlet:defineObjects />
+
+<%
+String viewMode = ParamUtil.getString(request, "viewMode");
+
+JournalArticleDisplay articleDisplay = (JournalArticleDisplay)request.getAttribute(WebKeys.JOURNAL_ARTICLE_DISPLAY);
+%>
+
+<c:if test="<%= !viewMode.equals(Constants.PRINT) %>">
+
+	<%
+	String[] availableLocales = articleDisplay.getAvailableLocales();
+
+	PortletURL currentURLObj = PortletURLUtil.getCurrent(liferayPortletRequest, liferayPortletResponse);
+
+	String currentURL = currentURLObj.toString();
+	%>
+
+	<c:if test="<%= availableLocales.length > 1 %>">
+		<div class="locale-actions user-tool-asset-addon-entry">
+			<liferay-ui:language formAction="<%= currentURL %>" languageId="<%= LanguageUtil.getLanguageId(request) %>" languageIds="<%= availableLocales %>" />
+		</div>
+	</c:if>
+</c:if>
